requirements.txt
aiomysql>=0.0.21
alembic>=1.6.5
asyncpg>=0.23.0
base58>=2.0.1
certifi>=2020.12.5
chardet>=3.0.4
idna>=2.8
more-itertools>=8.7.0
python-decouple>=3.4
redis>=3.5.3
scalecodec>=0.11.0
substrate-interface>=0.13.2
tqdm>=4.61.0
urllib3>=1.25.3
walrus>=0.8.2
xxhash>=2.0.2
